// back-end/src/roulette/roulette-bet-place.enum.ts
export enum RouletteBetPlace {
  // Singles
  ZERO = '0',
  _1 = '1',
  _2 = '2',
  _3 = '3',
  _4 = '4',
  _5 = '5',
  _6 = '6',
  _7 = '7',
  _8 = '8',
  _9 = '9',
  _10 = '10',
  _11 = '11',
  _12 = '12',
  _13 = '13',
  _14 = '14',
  _15 = '15',
  _16 = '16',
  _17 = '17',
  _18 = '18',
  _19 = '19',
  _20 = '20',
  _21 = '21',
  _22 = '22',
  _23 = '23',
  _24 = '24',
  _25 = '25',
  _26 = '26',
  _27 = '27',
  _28 = '28',
  _29 = '29',
  _30 = '30',
  _31 = '31',
  _32 = '32',
  _33 = '33',
  _34 = '34',
  _35 = '35',
  _36 = '36',

  // Splits
  _0_3 = '0 - 3',
  _3_6 = '3 - 6',
  _6_9 = '6 - 9',
  _9_12 = '9 - 12',
  _12_15 = '12 - 15',
  _15_18 = '15 - 18',
  _18_21 = '18 - 21',
  _21_24 = '21 - 24',
  _24_27 = '24 - 27',
  _27_30 = '27 - 30',
  _30_33 = '30 - 33',
  _33_36 = '33 - 36',
  _0_2 = '0 - 2',
  _2_5 = '2 - 5',
  _5_8 = '5 - 8',
  _8_11 = '8 - 11',
  _11_14 = '11 - 14',
  _14_17 = '14 - 17',
  _17_20 = '17 - 20',
  _20_23 = '20 - 23',
  _23_26 = '23 - 26',
  _26_29 = '26 - 29',
  _29_32 = '29 - 32',
  _32_35 = '32 - 35',
  _0_1 = '0 - 1',
  _1_4 = '1 - 4',
  _4_7 = '4 - 7',
  _7_10 = '7 - 10',
  _10_13 = '10 - 13',
  _13_16 = '13 - 16',
  _16_19 = '16 - 19',
  _19_22 = '19 - 22',
  _22_25 = '22 - 25',
  _25_28 = '25 - 28',
  _28_31 = '28 - 31',
  _31_34 = '31 - 34',
  _1_2 = '1 - 2',
  _2_3 = '2 - 3',
  _4_5 = '4 - 5',
  _5_6 = '5 - 6',
  _7_8 = '7 - 8',
  _8_9 = '8 - 9',
  _10_11 = '10 - 11',
  _11_12 = '11 - 12',
  _13_14 = '13 - 14',
  _14_15 = '14 - 15',
  _16_17 = '16 - 17',
  _17_18 = '17 - 18',
  _19_20 = '19 - 20',
  _20_21 = '20 - 21',
  _22_23 = '22 - 23',
  _23_24 = '23 - 24',
  _25_26 = '25 - 26',
  _26_27 = '26 - 27',
  _28_29 = '28 - 29',
  _29_30 = '29 - 30',
  _31_32 = '31 - 32',
  _32_33 = '32 - 33',
  _34_35 = '34 - 35',
  _35_36 = '35 - 36',

  // Streets
  _0_1_2 = '0 - 1 - 2',
  _0_2_3 = '0 - 2 - 3',
  _1_2_3 = '1 - 2 - 3',
  _4_5_6 = '4 - 5 - 6',
  _7_8_9 = '7 - 8 - 9',
  _10_11_12 = '10 - 11 - 12',
  _13_14_15 = '13 - 14 - 15',
  _16_17_18 = '16 - 17 - 18',
  _19_20_21 = '19 - 20 - 21',
  _22_23_24 = '22 - 23 - 24',
  _25_26_27 = '25 - 26 - 27',
  _28_29_30 = '28 - 29 - 30',
  _31_32_33 = '31 - 32 - 33',
  _34_35_36 = '34 - 35 - 36',

  // Corners
  _0_1_2_3 = '0 - 1 - 2 - 3',
  _1_2_4_5 = '1 - 2 - 4 - 5',
  _4_5_7_8 = '4 - 5 - 7 - 8',
  _7_8_10_11 = '7 - 8 - 10 - 11',
  _10_11_13_14 = '10 - 11 - 13 - 14',
  _13_14_16_17 = '13 - 14 - 16 - 17',
  _16_17_19_20 = '16 - 17 - 19 - 20',
  _19_20_22_23 = '19 - 20 - 22 - 23',
  _22_23_25_26 = '22 - 23 - 25 - 26',
  _25_26_28_29 = '25 - 26 - 28 - 29',
  _28_29_31_32 = '28 - 29 - 31 - 32',
  _31_32_34_35 = '31 - 32 - 34 - 35',
  _2_3_5_6 = '2 - 3 - 5 - 6',
  _5_6_8_9 = '5 - 6 - 8 - 9',
  _8_9_11_12 = '8 - 9 - 11 - 12',
  _11_12_14_15 = '11 - 12 - 14 - 15',
  _14_15_17_18 = '14 - 15 - 17 - 18',
  _17_18_20_21 = '17 - 18 - 20 - 21',
  _20_21_23_24 = '20 - 21 - 23 - 24',
  _23_24_26_27 = '23 - 24 - 26 - 27',
  _26_27_29_30 = '26 - 27 - 29 - 30',
  _29_30_32_33 = '29 - 30 - 32 - 33',
  _32_33_35_36 = '32 - 33 - 35 - 36',

  // Six Lines
  _1_2_3_4_5_6 = '1 - 2 - 3 - 4 - 5 - 6',
  _4_5_6_7_8_9 = '4 - 5 - 6 - 7 - 8 - 9',
  _7_8_9_10_11_12 = '7 - 8 - 9 - 10 - 11 - 12',
  _10_11_12_13_14_15 = '10 - 11 - 12 - 13 - 14 - 15',
  _13_14_15_16_17_18 = '13 - 14 - 15 - 16 - 17 - 18',
  _16_17_18_19_20_21 = '16 - 17 - 18 - 19 - 20 - 21',
  _19_20_21_22_23_24 = '19 - 20 - 21 - 22 - 23 - 24',
  _22_23_24_25_26_27 = '22 - 23 - 24 - 25 - 26 - 27',
  _25_26_27_28_29_30 = '25 - 26 - 27 - 28 - 29 - 30',
  _28_29_30_31_32_33 = '28 - 29 - 30 - 31 - 32 - 33',
  _31_32_33_34_35_36 = '31 - 32 - 33 - 34 - 35 - 36',

  // Dozens
  DOZEN_1_TO_12 = '1_to_12',
  DOZEN_13_TO_24 = '13_to_24',
  DOZEN_25_TO_36 = '25_to_36',

  // Columns (the 2:1 bets at bottom)
  COLUMN_1 = '2:1/1',
  COLUMN_2 = '2:1/2',
  COLUMN_3 = '2:1/3',

  // Outside Bets
  _1_TO_18 = '1_to_18',
  EVEN = 'Even',
  RED = 'Red',
  BLACK = 'Black',
  ODD = 'Odd',
  _19_TO_36 = '19_to_36',
}